Refactor iframe space app functionality to a separate class
Add a new action class for the iframe space app (extending the old space app action class) that must be extended in
order to use the iframe version. A meta tag in the SPA index.html is not needed anymore for specifying that the
iframe version should be used.
