--- spark-confluence/src/test/resources/com/k15t/spark/confluence/testiframeapp/index.html
@@ -1,8 +1,0 @@
-<html>
-    <head>
-        <meta name="decorator" content="spark.iframe-space-app"/>
-    </head>
-    <body>
-        <p id="should_be_stripped">This not be added to the space app template body when using iframes</p>
-    </body>
-</html>
