+++ spark-confluence/src/test/resources/com/k15t/spark/confluence/testspa/index.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <meta name="spark.app-base-url" content="noniframe/testapp/baseurl/">
+        <meta name="spark.app-base-url" content="spark/space/testapp/baseurl/">
         <script src="test-script.js"/>
         <link rel="stylesheet" href="test-styles.css"/>
     </head>
